requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.0.0
azure-storage-blob==12.16.0
azure-servicebus
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
csvkit==1.4.0
ExifRead==3.0.0
filelock==3.12.0
fonttools==4.48.1
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msrest==0.7.1
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
six==1.16.0
sympy==1.11.1
statistics==1.0.3.5
tifffile==2023.4.12
torch
torchvision==0.15.2
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
zipp==3.15.0
imutils